+++ src/hooks/DTEndLevelLayer.hpp
@@ -3,8 +3,12 @@
 using namespace geode::prelude;
 
 class $modify(DTEndLevelLayer, EndLevelLayer) {
+    struct Fields{
+        CCMenuItemSpriteExtra* skullBtn;
+    };
     
     void customSetup();
 
     void openDTLayer(CCObject*);
+    void onHideLayer(cocos2d::CCObject* sender);
 };